pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["TheBranchDriftCatalyst <djdanielsh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.3.0"
pydantic = "^2.8.2"
mypy = "^1.11.2"
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
markdown = "^3.7"


[tool.poetry.group.dev.dependencies]
django-stubs = "^5.0.4"
types-markdown = "^3.7.0.20240822"
black = "^24.8.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
factory-boy = "^3.3.1"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
